+++ src/main/java/com/edso/resume/api/service/DepartmentServiceImpl.java
@@ -250,21 +250,27 @@ DepartmentServiceImpl extends BaseService implements DepartmentService {
     public BaseResponse deleteDepartment(DeleteDepartmentRequest request) {
         BaseResponse response = new BaseResponse();
         try {
-            String id = request.getId();
-            Bson cond = Filters.eq(DbKeyConfig.ID, id);
-            Document idDocument = db.findOne(CollectionNameDefs.COLL_DEPARTMENT_COMPANY, cond);
+            Document department = db.findOne(CollectionNameDefs.COLL_PROFILE, Filters.eq(DbKeyConfig.DEPARTMENT_ID, request.getId()));
+            if (department == null) {
+                String id = request.getId();
+                Bson cond = Filters.eq(DbKeyConfig.ID, id);
+                Document idDocument = db.findOne(CollectionNameDefs.COLL_DEPARTMENT_COMPANY, cond);
 
-            if (idDocument == null) {
-                response.setFailed("Id này không tồn tại");
+                if (idDocument == null) {
+                    response.setFailed("Id này không tồn tại");
+                    return response;
+                }
+                db.delete(CollectionNameDefs.COLL_DEPARTMENT_COMPANY, cond);
+                response.setSuccess();
+                return response;
+            }else{
+                response.setFailed("Không thể xóa phòng ban này!");
                 return response;
             }
-
-            db.delete(CollectionNameDefs.COLL_DEPARTMENT_COMPANY, cond);
         } catch (Throwable e) {
             logger.error("Exception: ", e);
             response.setFailed("Hệ thống bận");
             return response;
         }
-        return new BaseResponse(0, "OK");
     }
 }
